+++ packages/apollo-engine-reporting/src/extension.ts
@@ -15,8 +15,11 @@ import {
   EngineReportingOptions,
   GenerateClientInfo,
   AddTraceArgs,
+  VariableValueModifier,
+  VariableValueModifierOptions,
 } from './agent';
 import { GraphQLRequestContext } from 'apollo-server-core/dist/requestPipelineAPI';
+import { operation } from 'retry';
 
 const clientNameHeaderKey = 'apollographql-client-name';
 const clientReferenceIdHeaderKey = 'apollographql-client-reference-id';
@@ -146,41 +149,25 @@ export class EngineReportingExtension<TContext = any>
       }
     }
 
-    if (this.options.privateVariables !== true && o.variables) {
-      // Note: we explicitly do *not* include the details.rawQuery field. The
-      // Engine web app currently does nothing with this other than store it in
-      // the database and offer it up via its GraphQL API, and sending it means
-      // that using calculateSignature to hide sensitive data in the query
-      // string is ineffective.
-      this.trace.details = new Trace.Details();
-      Object.keys(o.variables).forEach(name => {
-        if (
-          this.options.privateVariables &&
-          Array.isArray(this.options.privateVariables) &&
-          // We assume that most users will have only a few private variables,
-          // or will just set privateVariables to true; we can change this
-          // linear-time operation if it causes real performance issues.
-          this.options.privateVariables.includes(name)
-        ) {
-          // Special case for private variables. Note that this is a different
-          // representation from a variable containing the empty string, as that
-          // will be sent as '""'.
-          this.trace.details!.variablesJson![name] = '';
-        } else {
-          try {
-            this.trace.details!.variablesJson![name] = JSON.stringify(
-              o.variables![name],
-            );
-          } catch (e) {
-            // This probably means that the value contains a circular reference,
-            // causing `JSON.stringify()` to throw a TypeError:
-            // https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/JSON/stringify#Issue_with_JSON.stringify()_when_serializing_circular_references
-            this.trace.details!.variablesJson![name] = JSON.stringify(
-              '[Unable to convert value to JSON]',
-            );
-          }
-        }
-      });
+    if (o.variables) {
+      if (
+        this.options.maskVariableValues !== undefined ||
+        this.options.privateVariables == null
+      ) {
+        // The new option maskVariableValues will take precendence over the deprecated privateVariables option
+        this.trace.details = makeTraceDetails(
+          o.variables,
+          this.options.maskVariableValues,
+          o.queryString,
+        );
+      } else {
+        // DEPRECATED: privateVariables
+        // But keep supporting if it has been set.
+        this.trace.details = makeTraceDetailsLegacy(
+          o.variables,
+          this.options.privateVariables,
+        );
+      }
     }
 
     const clientInfo = this.generateClientInfo(o.requestContext);
@@ -454,4 +441,107 @@ function defaultGenerateClientInfo({ request }: GraphQLRequestContext) {
   } else {
     return {};
   }
+}
+
+// Creates trace details from request variables, given a specification for modifying
+// values of private or sensitive variables.
+// The details include the variables in the request, TODO(helen): and the modifier type.
+// If maskVariableValues is a bool or Array, it will act similarly to
+// to the to-be-deprecated options.privateVariables, except that the redacted variable
+// names will still be visible in the UI when 'true.'
+// If maskVariableValues is null, the policy will default to the 'true' case.
+export function makeTraceDetails(
+  variables: Record<string, any>,
+  maskVariableValues?:
+    | {
+        valueModifier: (
+          options: VariableValueModifierOptions,
+        ) => Record<string, any>;
+      }
+    | { privateVariableNames: Array<String> }
+    | { always: boolean },
+  operationString?: string,
+): Trace.Details {
+  const details = new Trace.Details();
+  const variablesToRecord = (() => {
+    if (maskVariableValues && 'valueModifier' in maskVariableValues) {
+      // Custom function to allow user to specify what variablesJson will look like
+      const originalKeys = Object.keys(variables);
+      const modifiedVariables = maskVariableValues.valueModifier({
+        variables: variables,
+        operationString: operationString,
+      });
+      return cleanModifiedVariables(originalKeys, modifiedVariables);
+    } else {
+      return variables;
+    }
+  })();
+
+  // Note: we explicitly do *not* include the details.rawQuery field. The
+  // Engine web app currently does nothing with this other than store it in
+  // the database and offer it up via its GraphQL API, and sending it means
+  // that using calculateSignature to hide sensitive data in the query
+  // string is ineffective.
+  Object.keys(variablesToRecord).forEach(name => {
+    if (
+      maskVariableValues == null ||
+      ('always' in maskVariableValues && maskVariableValues.always) ||
+      ('privateVariableNames' in maskVariableValues &&
+        // We assume that most users will have only a few private variables,
+        // or will just set privateVariables to true; we can change this
+        // linear-time operation if it causes real performance issues.
+        maskVariableValues.privateVariableNames.includes(name))
+    ) {
+      // Special case for private variables. Note that this is a different
+      // representation from a variable containing the empty string, as that
+      // will be sent as '""'.
+      details.variablesJson![name] = '';
+    } else {
+      try {
+        details.variablesJson![name] =
+          variablesToRecord[name] == null
+            ? ''
+            : JSON.stringify(variablesToRecord[name]);
+      } catch (e) {
+        // This probably means that the value contains a circular reference,
+        // causing `JSON.stringify()` to throw a TypeError:
+        // https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/JSON/stringify#Issue_with_JSON.stringify()_when_serializing_circular_references
+        details.variablesJson![name] = JSON.stringify(
+          '[Unable to convert value to JSON]',
+        );
+      }
+    }
+  });
+
+  return details;
+}
+
+// Creates trace details when privateVariables [DEPRECATED] was set.
+// Wraps privateVariable into a format that can be passed into makeTraceDetails(),
+// which also handles the new replacement option maskVariableValues.
+export function makeTraceDetailsLegacy(
+  variables: Record<string, any>,
+  privateVariables: Array<String> | boolean,
+): Trace.Details | undefined {
+  const newArguments = Array.isArray(privateVariables)
+    ? {
+        privateVariableNames: privateVariables,
+      }
+    : {
+        always: privateVariables,
+      };
+  return makeTraceDetails(variables, newArguments);
+}
+
+// Helper for makeTraceDetails() to enforce that the keys of a modified 'variables'
+// matches that of the original 'variables'
+function cleanModifiedVariables(
+  originalKeys: Array<string>,
+  modifiedVariables: Record<string, any>,
+): Record<string, any> {
+  let cleanedVariables: Record<string, any> = {};
+  originalKeys.forEach(name => {
+    cleanedVariables[name] = modifiedVariables[name];
+  });
+  return cleanedVariables;
 }